Included kiraBot
Added kiraBot to the repo to show how you can use the API. Please note: you will need a .env file for the API and if you want to use kiraBot you will also need a users.json file (because I never bothered to make it be auto-created) that might (don't remember) require the contents of ' {} '. Honestly, if you're using this there's a fair chance you can figure it out just looking at it.