Added files and combo tags
Users can now specify if they only want to be tagged for tag combos. for example, a user may only want to be pinged if something has both the "monster_girl" AND "animal_ear_fluff" at the same time, and not just one of them. Also fixed a bug where the bot was unable to ping users and manage repost detection/notification.